refactor: ChatType.ts 파일 생성, 중복사용 타입 통합
