+++ src/features/chat/MockChatData.ts
@@ -1,24 +1,4 @@
-export interface ProductInfo {
-  image: string;
-  title: string;
-  price: string;
-}
-
-export interface ChatMessage {
-  message: string;
-  time: string;
-  position: 'left' | 'right';
-}
-
-export interface ChatItem {
-  roomId: string;
-  userName: string;
-  time: string;
-  lastMessage: string;
-  userAvatar: string;
-  product: ProductInfo;
-  messages: ChatMessage[];
-}
+import type { ChatItem } from 'src/types/ChatType';
 
 export const mockChatList: ChatItem[] = [
   {
